add ConfigUtils class, CompletionMetadataTest class
